Add files via upload
REPORT 1 -
The report gives a high level performance of Managers. The report can be mostly used by VP of the company. The report shows:
1.	KPI’s like Sales, Profit and Number of Orders
2.	Sales and Profit by Month for selected Year
3.	Top 5 Customer by Profit Ratio

The number of visuals used in the report are 5. Among  these 5, 1 is a Custom visual and rest are In built Power BI visuals.
 
The extra columns and measures used in the report are as follows:
1.	Year derived from Order date
2.	Month Year derived from Order date
3.	Month sort used for sorting 
4.	Profit Ratio

REPORT 2 - 
The report gives a the analysis of Product SubCategory across Customer Segment and Product Category for the selected year. The report can be mostly used by the Product Manager to study the performance of Product SubCategory. The report shows:
1.	Count of Product Subcategory
2.	Sales and Profit by Month for selected Year 

The number of visuals used in the report are 3. Among  these 3, 2 are Custom visual and rest are In built Power BI visuals.
 
The extra columns and measures used in the report are as follows:
1.	Year derived from Order date
2.	Month Year derived from Order date
3.	Month sort used for sorting 

REPORT 3 - 
The report gives a the detailed view of Product Category, Product SubCategory and Products for the selected year. The report shows:
1.	Sales of Product 
2.	Sales and Profit by Product Category
3.	Sale and Profit by Product Sub-Category
4.	Top Performing Products
5.	All Prodcut Details

The number of visuals used in the report are 5. Among  these 5, 1 is Custom visual and rest are In built Power BI visuals.
 
The extra columns and measures used in the report are as follows:
1.	Year derived from Order date
2.	Profit Ratio